WW-5293 Split XmlConfigurationProvider into XmlDocConfigurationProvider Part 2
